adding parsers and label
This commit will add an ability to add a custom metadata label
(eg a doi) via the command line or internal function. We are also
adding a primary parser (gitlab) and a secondary parser (zenodo)
and the starting idea for a scraper, which will be able to automatically
discover new software repositories.

Signed-off-by: vsoch <vsochat@stanford.edu>